refactored error handling and validators
