+++ pydantic_marc/field_validators.py
@@ -1,15 +1,7 @@
-"""
-Error handling functions for MARC field validation.
+"""Custom validation functions for `ControlField`, `DataField`, and `MarcRecord` models.
 
-This module defines specialized error-checking logic for different components of
-a MARC record (ie. `ControlField.data`, `DataField.indicators`, `DataField.subfields`,
-`MarcRecord.fields`). Each function evaluates its input against a set of rules
-defined in a `Rule` or `RuleSet` and returns a list of structured error details if
-any violations are found.
-
-Functions in this module are invoked by custom validators during model validation
-and return lists of `InitErrorDetails` objects to support Pydantic-compatible error
-reporting.
+The validator functions in this module are used as either after or wrap validators
+depending on the field and model.
 """
 
 from __future__ import annotations
@@ -18,7 +10,7 @@ import json
 import os
 from collections import Counter
 from functools import lru_cache
-from typing import TYPE_CHECKING, Any, Sequence, Union
+from typing import TYPE_CHECKING, Any, Callable, Sequence, Union
 
 from pydantic import ValidationInfo
 from pydantic_core import InitErrorDetails
@@ -27,17 +19,13 @@ from .errors import (
     ControlFieldLength,
     InvalidFixedField,
     InvalidIndicator,
-    InvalidLeader,
     InvalidSubfield,
-    MissingRequiredField,
-    MultipleMainEntryValues,
-    NonRepeatableField,
     NonRepeatableSubfield,
+    raise_validation_errors,
 )
 
 if TYPE_CHECKING:  # pragma: no cover
     from .components import PydanticIndicators, PydanticSubfield
-    from .marc_rules import Rule
 
 
 @lru_cache
@@ -50,8 +38,70 @@ def marc_codes() -> dict[str, Any]:
     return rules
 
 
+def marc_field_validator(error_checker_func: Callable) -> Callable:
+    """
+    Wrapper function to run field-level validation for a MARC field using
+    rules from the model.
+
+    This function looks up a validation function based on the field name and applies
+    it if a corresponding rule and validator are found. If no rule or validator is
+    found, the input `data` is returned unchanged. If validation errors are found,
+    they are raised using `raise_validation_errors`.
+
+    Args:
+        data: the data passed to the field being validated
+        info: A `ValidationInfo` object.
+
+    Returns:
+        The validated field data, or raises a `ValidationError` if rules are violated.
+    """
+
+    def decorator(func: Callable) -> Callable:
+        def wrapper(data: Any, info: ValidationInfo) -> Any:
+            rule = info.data.get("rules", None)
+            if not rule:
+                return data
+            errors = error_checker_func(
+                rule=rule.model_dump(), data=data, tag=info.data["tag"]
+            )
+            return raise_validation_errors(errors=errors, data=data)
+
+        return wrapper
+
+    return decorator
+
+
+def get_control_field_length_errors(
+    rule: dict[str, Any], data: str, tag: str
+) -> list[InitErrorDetails]:
+    """
+    Validate the length of a control field's `data` string against the expected rule.
+
+    If the `data` string length does not match the expected length specified in
+    the rule, a `ControlFieldLength` error is returned.
+
+    Args:
+        rule: the `Rule` object defining the expected length and values for the field.
+        data: data passed to the `ControlField.data` attribute.
+        tag: the MARC field tag being validated.
+
+    Returns:
+
+        A list of `MarcCustomError` objects.
+    """
+    errors: list[InitErrorDetails] = []
+    valid_length = rule.get("length")
+    if not valid_length:
+        return errors
+    match = len(data) == valid_length
+    if match is False:
+        error = ControlFieldLength({"tag": tag, "valid": valid_length, "input": data})
+        errors.append(error.error_details)
+    return errors
+
+
 def get_control_field_value_errors(
-    rule: Rule, data: str, tag: str
+    rule: dict[str, Any], data: str, tag: str
 ) -> list[InitErrorDetails]:
     """
     Validate the values of each character of a control field's `data` string
@@ -72,7 +122,7 @@ def get_control_field_value_errors(
         A list of `MarcCustomError` objects.
     """
     errors: list[InitErrorDetails] = []
-    value_rules = rule.values
+    value_rules = rule.get("values")
     if value_rules:
         data_dict = {f"{i:02d}": char for i, char in enumerate(data)}
         for position in value_rules.keys():
@@ -112,37 +162,8 @@ def get_control_field_value_errors(
     return errors
 
 
-def get_control_field_length_errors(
-    rule: Rule, data: str, tag: str
-) -> list[InitErrorDetails]:
-    """
-    Validate the length of a control field's `data` string against the expected rule.
-
-    If the `data` string length does not match the expected length specified in
-    the rule, a `ControlFieldLength` error is returned.
-
-    Args:
-        rule: the `Rule` object defining the expected length and values for the field.
-        data: data passed to the `ControlField.data` attribute.
-        tag: the MARC field tag being validated.
-
-    Returns:
-
-        A list of `MarcCustomError` objects.
-    """
-    errors: list[InitErrorDetails] = []
-    valid_length = rule.length
-    if not valid_length:
-        return errors
-    match = len(data) == valid_length
-    if match is False:
-        error = ControlFieldLength({"tag": tag, "valid": valid_length, "input": data})
-        errors.append(error.error_details)
-    return errors
-
-
 def get_indicator_errors(
-    rule: Rule, data: Union[PydanticIndicators, Sequence], tag: str
+    rule: dict[str, Any], data: Union[PydanticIndicators, Sequence], tag: str
 ) -> list[InitErrorDetails]:
     """
     Validate the indicator values of a `DataField` against the allowed values in a rule.
@@ -162,7 +183,7 @@ def get_indicator_errors(
     errors: list[InitErrorDetails] = []
     for n, indicator in enumerate(data):
         ind = f"ind{n + 1}"
-        valid_inds = getattr(rule, ind)
+        valid_inds = rule.get(ind, "")
         if data[n] not in valid_inds:
             error_data = {"loc": (tag, ind), "input": indicator, "valid": valid_inds}
             errors.append(InvalidIndicator(error_data).error_details)
@@ -170,7 +191,7 @@ def get_indicator_errors(
 
 
 def get_subfield_errors(
-    rule: Rule, data: list[PydanticSubfield], tag: str
+    rule: dict[str, Any], data: list[PydanticSubfield], tag: str
 ) -> list[InitErrorDetails]:
     """
     Validate the subfields in a `DataField` against the allowed and repeatable values
@@ -194,14 +215,15 @@ def get_subfield_errors(
         A list of `MarcCustomError` objects.
     """
     errors: list[InitErrorDetails] = []
-    if not rule.subfields:
+    sub_rules = rule.get("subfields")
+    if not sub_rules:
         return errors
 
     sub_codes: Counter = Counter([sub.code for sub in data])
     deduped_sub_codes = set(sub_codes.elements())
 
-    valid_sub_codes = rule.subfields.get("valid", [])
-    nr_sub_codes = rule.subfields.get("non_repeatable", [])
+    valid_sub_codes = sub_rules.get("valid", [])
+    nr_sub_codes = sub_rules.get("non_repeatable", [])
 
     for code in deduped_sub_codes:
         error_data = {"loc": (tag, code), "input": [i for i in data if i.code == code]}
@@ -209,79 +231,4 @@ def get_subfield_errors(
             errors.append(NonRepeatableSubfield(error_data).error_details)
         elif valid_sub_codes and code not in valid_sub_codes:
             errors.append(InvalidSubfield(error_data).error_details)
-    return errors
-
-
-def get_marc_field_errors(
-    data: list[Any], info: ValidationInfo
-) -> list[InitErrorDetails]:
-    """
-    Validate rules across all fields in a `MarcRecord`.
-
-    This includes:
-    - Checking that non-repeatable fields appear no more than once.
-    - Confirming that required fields are present.
-    - Ensuring that only one main entry field (1XX tag) exists.
-
-    This function is called before validation of the `MarcRecord` model within the
-    `WrapValidator` on the `MarcRecord.fields` attribute.
-
-    Args:
-        data: A list of fields from a `MarcRecord`.
-        info: A `ValidationInfo` context used to extract applicable rules.
-
-    Returns:
-
-        A list of `MarcCustomError` objects.
-    """
-    errors: list[InitErrorDetails] = []
-    rules = info.data["rules"]
-    if not rules:
-        return errors
-    tag_counts = Counter([i["tag"] for i in data])
-    main_entries = [i for i in tag_counts.elements() if i.startswith("1")]
-
-    for tag in rules.non_repeatable_fields:
-        if tag_counts[tag] > 1:
-            errors.append(NonRepeatableField({"input": tag}).error_details)
-    for tag in rules.required_fields:
-        if tag not in tag_counts.elements():
-            errors.append(MissingRequiredField({"input": tag}).error_details)
-    if len(main_entries) > 1:
-        errors.append(MultipleMainEntryValues({"input": main_entries}).error_details)
-    return errors
-
-
-def get_leader_errors(data: str, info: ValidationInfo) -> list[InitErrorDetails]:
-    """
-    Validate each character in a string against the allowed values each byte in a
-    MARC leader.
-
-    If the value does not match the rules for the leader, an `InvalidLeader`
-    error will be added to the list of errors and returned.
-
-    Args:
-        data: A string passed to the `MarcRecord.leader` attribute.
-        info: A `ValidationInfo` context used to extract applicable rules.
-
-    Returns:
-
-        A list of `MarcCustomError` objects.
-    """
-    errors: list[InitErrorDetails] = []
-    rules = info.data["rules"]
-    if not rules or not rules.rules:
-        return errors
-    rule = rules.rules.get("LDR")
-    if not rule or not rule.values:
-        return errors
-    for i, c in enumerate(data):
-        position = str(i).zfill(2)
-        valid = rule.values.get(f"{position}", [])
-        if c not in valid:
-            errors.append(
-                InvalidLeader(
-                    {"input": c, "loc": f"{position}", "valid": valid}
-                ).error_details
-            )
     return errors
