+++ pydantic_marc/fields.py
@@ -13,16 +13,37 @@ from __future__ import annotations
 
 from typing import Annotated, Any, Literal, Sequence, Union
 
-from pydantic import AfterValidator, BaseModel, Field, model_serializer
+from pydantic import AfterValidator, BaseModel, Field, ValidationInfo, model_serializer
 
 from .components import PydanticIndicators, PydanticSubfield
-from .marc_rules import Rule
-from .validators import (
-    validate_indicators,
-    validate_length,
-    validate_subfields,
-    validate_values,
+from .field_validators import (
+    get_control_field_length_errors,
+    get_control_field_value_errors,
+    get_indicator_errors,
+    get_subfield_errors,
+    marc_field_validator,
 )
+from .marc_rules import Rule
+
+
+@marc_field_validator(get_control_field_length_errors)
+def validate_length(data: str, info: ValidationInfo) -> None: ...  # pragma: no cover
+
+
+@marc_field_validator(get_control_field_value_errors)
+def validate_values(data: str, info: ValidationInfo) -> None: ...  # pragma: no cover
+
+
+@marc_field_validator(get_indicator_errors)
+def validate_indicators(
+    data: Union[PydanticIndicators, Sequence], info: ValidationInfo
+) -> None: ...  # pragma: no cover
+
+
+@marc_field_validator(get_subfield_errors)
+def validate_subfields(
+    data: list[PydanticSubfield], info: ValidationInfo
+) -> None: ...  # pragma: no cover
 
 
 class ControlField(BaseModel, arbitrary_types_allowed=True, from_attributes=True):
