--- pydantic_marc/utils.py
@@ -1,118 +1,0 @@
-"""
-Utility functions for MARC record validation.
-
-This module provides helper functions used during the validation of MARC records,
-specifically for handling validation rules and validation errors in a structured
-way.
-
-Functions:
-
-add_rules_to_pymarc_fields:
-    Adds rule metadata from a `RuleSet` to MARC fields.
-handle_errors:
-    Wraps validators to capture and normalize Pydantic validation errors.
-raise_validation_errors:
-    Raises a `ValidationError` from a list of error details.
-"""
-
-from __future__ import annotations
-
-from typing import Any, Callable
-
-from pydantic import ValidationError, ValidationInfo
-from pydantic_core import InitErrorDetails
-
-from .errors import MarcCustomError
-
-
-def add_rules_to_pymarc_fields(data: list[Any], info: ValidationInfo) -> list[Any]:
-    """
-    Add rules to each MARC field in a record.
-
-    This function creates a `RuleSet` from the provided `ValidationInfo` and uses it
-    to add each field in the `data` list with the appropriate `Rule` based on the
-    field's tag. The function constructs a list of dicts representing the fields,
-    each enriched with its tag, rule, and either control field data or
-    indicators/subfields.
-
-    Args:
-        data:
-            a list of `pymarc.Field` objects or similarly structured data.
-            The data is assumed to contain `tag`, and `control_field` attributes
-            and, depending on the field, `indicators`, `subfields`, and/or `data`
-            attributes.
-        info:
-            the `ValidationInfo` object provided during model validation.
-
-    Returns:
-        a list of dictionaries where each item represents a MARC field and its
-        associated `Rule`.
-    """
-    rule_set = info.data["rules"]
-    rules = rule_set.rules if rule_set else {}
-    field_list = []
-
-    for field in data:
-        rule = rules.get(field.tag, None)
-        if field.tag == "007" and rule:
-            rule = rule.get(field.data[0], {})
-        field_dict = {"rules": rule, "tag": field.tag}
-        if getattr(field, "control_field", field.is_control_field()) is True:
-            field_dict["data"] = field.data
-        else:
-            field_dict["indicators"] = field.indicators
-            field_dict["subfields"] = field.subfields
-        field_list.append(field_dict)
-    return field_list
-
-
-def handle_errors(
-    validator: Callable, data: Any, info: ValidationInfo
-) -> tuple[Any, list[InitErrorDetails]]:
-    """
-    Execute a validator function and convert any errors raised to custom errors.
-
-    This function wraps the execution of a validator, catching `ValidationError`
-    exceptions and converting each error into a `MarcCustomError`. It returns the
-    original or validated data along with a list of `InitErrorDetails` used for raising
-    or aggregating errors later.
-
-    Args:
-        validator: a callable validator function that takes `data` and `info`.
-        data: the input data as a list to be validated.
-        info: the `ValidationInfo` used during model validation.
-
-    Returns:
-        a tuple containing the validated data (or original data if invalid), and a
-        list of Pydantic-compatible error details.
-    """
-    try:
-        return validator(data, info), []
-    except ValidationError as exc:
-        errors = [MarcCustomError(e["type"], e["msg"], e["ctx"]) for e in exc.errors()]
-        return data, [i.error_details for i in errors]
-
-
-def raise_validation_errors(errors: list[InitErrorDetails], data: Any) -> Any:
-    """
-    Raise a `ValidationError` if any collected error details exist.
-
-    This function takes a list of Pydantic `InitErrorDetails` objects and raises a
-    `ValidationError` if the list is not empty. Otherwise, it returns the validated
-    data.
-
-    Args:
-        errors: a list of error details to raise, if any.
-        data: the data object being validated (used to name the error context).
-
-    Returns:
-        The input data, validated, if no errors were raised.
-
-    Raises:
-        `ValidationError` from the collected `errors`.
-    """
-    if errors:
-        raise ValidationError.from_exception_data(
-            title=data.__class__.__name__, line_errors=errors
-        )
-    return data
